The WIP Package builder can now build certain packages automatically
Note: Not every package works yet, for example packages that specify a specific version or packages that are provided by a package with a different name. It assumes it's an AUR package since it can't find it in the official repositories, and it doesn't execute the AUR check due to a bug in that function that causes it to crash.

Working example: "system-monitoring-center"
Not working example: "protontricks" (Tries to install "python>=3.3")